Make proposal cards use CardMCell
