+++ decidim-core/app/cells/decidim/card_m/author.erb
@@ -1,0 +1,3 @@
+<div class="card__author">
+  <%= card_for present(model).author, context: { extra_classes: ["author-data--small"], from: from_context } %>
+</div>
