+++ decidim-core/app/cells/decidim/card_m/badge.erb
@@ -1,0 +1,1 @@
+<%= content_tag :span, badge_name, class: badge_classes %>
