+++ decidim-core/app/cells/decidim/card_m/header.erb
@@ -4,4 +4,6 @@
       <%= title %>
     </h5>
   <% end %>
+
+  <%= render :author if has_author? %>
 </div>
