+++ decidim-core/app/cells/decidim/card_m/show.erb
@@ -1,18 +1,20 @@
 <div class="column" id="<%= dom_id(model) %>">
-  <article class="card card--<%= dom_class(model) %>">
+  <article class="card <%= card_classes %>">
     <div class="card__content">
       <%= render :header %>
 
       <div class="card__text">
         <div class="card__text--paragraph">
+          <%= render :badge if has_badge? %>
           <%= description %>
-          <%= render :tags %>
         </div>
       </div>
+
+      <%= render :tags %>
     </div>
 
     <%= render :data %>
     <%= render :status %>
-    <%= render :footer if shows_footer? %>
+    <%= render :footer %>
   </article>
 </div>
