--- decidim-proposals/app/cells/decidim/proposals/proposal_m/badge.erb
@@ -1,3 +1,0 @@
-<% if answered? || withdrawn? %>
-  <%= content_tag :span, state_name, class: badge_classes %>
-<% end %>
