--- decidim-proposals/app/cells/decidim/proposals/proposal_m/header.erb
@@ -1,11 +1,0 @@
-<div class="card__header">
-  <%= link_to resource_path, class: "card__link" do %>
-    <h5 class="card__title">
-      <%= title %>
-    </h5>
-  <% end %>
-
-  <div class="card__author">
-    <%= card_for present(model).author, context: { extra_classes: ["author-data--small"], from: from_context } %>
-  </div>
-</div>
