--- decidim-proposals/app/cells/decidim/proposals/proposal_m/show.erb
@@ -1,19 +1,0 @@
-<div class="column" id="<%= dom_id(model) %>">
-  <article class="card card--proposal <%= badge_state_css %>">
-    <div class="card__content">
-      <%= render :header %>
-
-      <div class="card__text">
-        <div class="card__text--paragraph">
-          <%= render :badge %>
-          <%= body %>
-        </div>
-      </div>
-
-      <%= cell "decidim/tags", model, context: {extra_classes: ["tags--proposal"]} %>
-    </div>
-
-    <%= render :status %>
-    <%= render :footer %>
-  </article>
-</div>
